refactor: move Tooltip component to core directory
